revno: 14078 message:   minor change to event controller to allow updating of single value
committer: Abyot Asalefew Gizaw <abyota@gmail.com>
branch nick: dhis2
timestamp: Thu 2014-02-20 11:21:30 +0100

+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/events/event/AbstractEventService.java
@@ -434,7 +434,7 @@ public abstract class AbstractEventService
     // -------------------------------------------------------------------------
 
     @Override
-    public void updateEvent( Event event )
+    public void updateEvent( Event event, boolean singleValue )
     {
         ProgramStageInstance programStageInstance = programStageInstanceService.getProgramStageInstance( event
             .getEvent() );
@@ -518,13 +518,17 @@ public abstract class AbstractEventService
                     value.getProvidedElsewhere() );
             }
         }
-
-        for ( TrackedEntityDataValue value : dataValues )
+        
+        if( !singleValue )
         {
-            dataValueService.deleteTrackedEntityDataValue( value );
+            for ( TrackedEntityDataValue value : dataValues )
+            {
+                dataValueService.deleteTrackedEntityDataValue( value );
+            }   
         }
+        
     }
-
+    
     // -------------------------------------------------------------------------
     // DELETE
     // -------------------------------------------------------------------------
